+++ packages/components/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @scalar/components
 
+## 0.14.23
+
+### Patch Changes
+
+- Updated dependencies [46bfd2f]
+  - @scalar/oas-utils@0.4.18
+
+## 0.14.22
+
+### Patch Changes
+
+- dd00357: feat(components): add support for custom combobox items
+- Updated dependencies [46ebe6e]
+  - @scalar/oas-utils@0.4.17
+
 ## 0.14.21
 
 ### Patch Changes