+++ packages/openapi-to-markdown/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @scalar/openapi-to-markdown
 
+## 0.2.26
+
+### Patch Changes
+
+- Updated dependencies [46bfd2f]
+  - @scalar/oas-utils@0.4.18
+  - @scalar/components@0.14.23
+
+## 0.2.25
+
+### Patch Changes
+
+- Updated dependencies [952bde2]
+- Updated dependencies [46ebe6e]
+- Updated dependencies [dd00357]
+- Updated dependencies [2888e18]
+  - @scalar/openapi-parser@0.19.0
+  - @scalar/oas-utils@0.4.17
+  - @scalar/components@0.14.22
+
 ## 0.2.24
 
 ### Patch Changes