+++ packages/postman-to-openapi/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @scalar/postman-to-openapi
 
+## 0.3.21
+
+### Patch Changes
+
+- Updated dependencies [46bfd2f]
+  - @scalar/oas-utils@0.4.18
+
+## 0.3.20
+
+### Patch Changes
+
+- Updated dependencies [46ebe6e]
+  - @scalar/oas-utils@0.4.17
+
 ## 0.3.19
 
 ### Patch Changes