+++ packages/use-codemirror/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @scalar/use-codemirror
 
+## 0.12.24
+
+### Patch Changes
+
+- @scalar/components@0.14.23
+
+## 0.12.23
+
+### Patch Changes
+
+- Updated dependencies [dd00357]
+  - @scalar/components@0.14.22
+
 ## 0.12.22
 
 ### Patch Changes